box damage and player kill

+++ JumpTheGun-Conversion/Assets/Scripts/Components/Health.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 [GenerateAuthoringComponent]
 public struct Health : IComponentData
 {
-    public int Value;
+    public float Value;
 }
